+++ src/paddlefleet/transformer/csa_attention.py
@@ -46,6 +46,11 @@ from paddlefleet.transformer.dsa_attention import (
     rotate_activation,
 )
 
+from paddlefleet.transformer.utils import (
+    get_doc_lens,
+    get_doc_starts,
+    )
+
 if TYPE_CHECKING:
     from paddlefleet.process_groups_config import ProcessGroupCollection
     from paddlefleet.transformer.enums import AttnMaskType
@@ -54,48 +59,6 @@ if TYPE_CHECKING:
 # ---------------------------------------------------------------------------
 # Helper functions for index computation
 # ---------------------------------------------------------------------------
-
-
-def get_doc_lens(startend_row_indices: Tensor) -> Tensor:
-    """Derive document lengths from startend_row_indices.
-
-    Args:
-        startend_row_indices: [batch_size, h, seqlen, 1] tensor where
-            each value is the end boundary (exclusive) of the document that
-            position belongs to.
-
-    Returns:
-        doc_lens: [n_docs] int32 tensor of document lengths.
-    """
-    mask = startend_row_indices.flatten().cast("int64")
-    seqlen = mask.shape[0]
-    positions = paddle.arange(seqlen, dtype="int64")
-
-    is_boundary = paddle.zeros([seqlen], dtype="bool")
-    is_boundary[0] = True
-    is_boundary[1:] = (positions[1:] == mask[:-1]) & (mask[1:] != mask[:-1])
-
-    boundary_indices = paddle.nonzero(is_boundary).flatten()
-    doc_ends = mask[boundary_indices]
-    doc_lens = (doc_ends - boundary_indices).cast("int32")
-    return doc_lens
-
-
-def get_doc_starts(doc_lens: Tensor) -> Tensor:
-    """Compute document start positions from document lengths.
-
-    Args:
-        doc_lens: [n_docs] tensor of document lengths.
-
-    Returns:
-        doc_starts: [n_docs] int32 tensor of cumulative start positions.
-    """
-    lens = doc_lens.flatten().cast("int32")
-    cum = paddle.cumsum(lens, axis=0)
-    starts = paddle.zeros_like(cum)
-    if cum.shape[0] > 1:
-        starts[1:] = cum[:-1]
-    return starts
 
 
 def get_cutoff_doc_lens(doc_lens: Tensor, ratio: int) -> Tensor:
